requirements.txt
pandas
numpy
scikit-learn
torch
transformers
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
nltk
matplotlib
joblib
python-multipart
httpx
openai
email-validator
